# openapi.yaml
openapi: 3.0.3
info:
  title: računAI API
  version: 1.0.0
  description: Tenant JWT REST ugovor za auth, documents, banking, partners, finance,
    assets, purchasing i tax. Fiscal/intermediary i AS4 nisu dio ove sheme. openapi.yaml
    je generirani artefakt — ne uređivati ručno.
paths:
...
  /api/tax/pdv-s/periods/{period}/:
    get:
      operationId: tax_pdv_s_period_retrieve
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvSPeriod'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/tax/pdv-s/periods/{period}/submit/:
    post:
      operationId: tax_pdv_s_period_submit
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PdvSSubmitRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PdvSSubmitRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PdvSSubmitRequestRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResult'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
  /api/tax/pdv-s/periods/{period}/xml/:
    get:
      operationId: tax_pdv_s_period_xml
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: Unsigned Obrazac PDV-S XML
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/tax/pdv/periods/:
    get:
      operationId: tax_pdv_periods_list
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvPeriodList'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/tax/pdv/periods/{period}/:
    get:
      operationId: tax_pdv_period_retrieve
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvPeriodWorkspace'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/tax/pdv/periods/{period}/boxes/:
    get:
      operationId: tax_pdv_period_boxes
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvBoxes'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/tax/pdv/periods/{period}/correction/:
    post:
      operationId: tax_pdv_period_correction
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvDraft'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
  /api/tax/pdv/periods/{period}/draft/:
    post:
      operationId: tax_pdv_period_draft
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvDraft'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
  /api/tax/pdv/periods/{period}/ledger/:
    post:
      operationId: tax_pdv_period_ledger
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdvLedgerRebuild'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
  /api/tax/pdv/periods/{period}/submit/:
    post:
      operationId: tax_pdv_period_submit
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PdvSubmitRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PdvSubmitRequestRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/PdvSubmitRequestRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResult'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
  /api/tax/pdv/periods/{period}/xml/:
    get:
      operationId: tax_pdv_period_xml
      parameters:
      - in: path
        name: period
        schema:
          type: string
        required: true
      tags:
      - tax
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: Unsigned Obrazac PDV XML
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Konflikt (idempotency / match target taken)
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      description: 'DRF-style error body: ``{"detail": ...}`` where detail may be
        str or object.'
      properties:
        detail: {}
      required:
      - detail
...
    PdvPeriodList:
      type: object
      properties:
        count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/PdvPeriod'
      required:
      - count
      - results
...
    PdvSPeriod:
      type: object
      properties:
        period:
          type: string
        schema_version:
          type: string
        row_count:
          type: integer
        total_goods:
          type: string
          description: Decimal as string, e.g. "1100.00"
        total_services:
          type: string
          description: Decimal as string, e.g. "1100.00"
        rows:
          type: array
          items:
            $ref: '#/components/schemas/PdvSRow'
        event_uuid:
          type: string
          format: uuid
          nullable: true
        current_submission:
          allOf:
          - $ref: '#/components/schemas/PdvSSubmission'
          nullable: true
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/PdvSSubmission'
      required:
      - current_submission
      - event_uuid
      - period
      - row_count
      - rows
      - schema_version
      - submissions
      - total_goods
      - total_services
...
    PdvSSubmitRequestRequest:
      type: object
      properties:
        eporezna_identifier:
          type: string
          format: uuid
        submitted_at:
          type: string
          format: date-time
      required:
      - eporezna_identifier
      - submitted_at
...
    SubmissionResult:
      type: object
      properties:
        event_uuid:
          type: string
          format: uuid
        external_identifier:
          type: string
          format: uuid
        submitted_at:
          type: string
        has_confirmation:
          type: boolean
      required:
      - event_uuid
      - external_identifier
      - has_confirmation
      - submitted_at
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...